Improve tidyup error message
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@759808 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/compress/AbstractTestCase.java
@@ -70,8 +70,9 @@ public abstract class AbstractTestCase extends TestCase {
         rmdir(dir);
         rmdir(resultDir);
         dir = resultDir = null;
-        if (archive != null) {
+        if (archive != null && archive.exists()) {
             if (!archive.delete()){
+                // Note: this exception won't be shown if the test has already failed
                 throw new Exception("Could not delete "+archive.getPath());
             }
         }
